Moved GameStateFlags to its own header file
