+++ src/OpenLoco/src/GameState.h
@@ -11,18 +11,10 @@
 #include "ScenarioObjective.h"
 #include "Station.h"
 #include "Town.h"
-#include <OpenLoco/Core/EnumFlags.hpp>
 
 namespace OpenLoco
 {
-    enum class GameStateFlags : uint32_t
-    {
-        none = 0U,
-        tileManagerLoaded = (1U << 0),
-        unk2 = (1U << 1),
-        preferredOwnerName = (1U << 2),
-    };
-    OPENLOCO_ENABLE_ENUM_OPERATORS(GameStateFlags);
+    enum class GameStateFlags : uint32_t;
 
 #pragma pack(push, 1)
     struct GameState
