+++ src/OpenLoco/src/TownManager.cpp
@@ -2,6 +2,7 @@
 #include "CompanyManager.h"
 #include "Game.h"
 #include "GameState.h"
+#include "GameStateFlags.h"
 #include "Interop/Interop.hpp"
 #include "Map/BuildingElement.h"
 #include "Map/TileLoop.hpp"
