+++ src/OpenLoco/src/Ui.cpp
@@ -1,5 +1,5 @@
 #include "Drawing/SoftwareDrawingEngine.h"
-#include "GameState.h"
+#include "GameStateFlags.h"
 #include "Scenario.h"
 #include "Ui/Cursor.h"
 #include <algorithm>
